% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gentry}
\alias{gentry}
\title{Gentry forest transect dataset as a phyloseq object}
\format{
A \code{\link[phyloseq]{phyloseq-class}} object containing:
\describe{
\item{OTU table}{Species abundance counts, with rows as samples (sites) and columns as species identified by their trinomial names.}
\item{phylogenetic tree}{A phylo object representing the evolutionary relationships among species in the dataset.}
\item{sample data}{Metadata for each forest plot site, including geographic coordinates and other site attributes.}
\item{taxonomy table}{Taxonomic classifications (family, genus, species) matched using the \code{\link[taxonlookup]{taxonlookup}} package.}
}
}
\source{
Raw data originally from:
\url{https://github.com/zdealveindy/anadat-r/tree/master/data}

Phylogenetic tree and nodes from:
\url{https://github.com/jinyizju/S.PhyloMaker}

Taxonomic lookups performed using the \code{\link[taxonlookup]{taxonlookup}} package.
}
\usage{
gentry
}
\description{
A dataset containing species abundance data, phylogenetic tree, sample metadata, and taxonomy
information for forest plots from the Gentry dataset. This dataset has been processed to
aggregate species counts by unique trinomial names, mapped onto a phylogenetic tree using
the S.PhyloMaker tool, and organized into a \code{\link[phyloseq]{phyloseq-class}} object.
}
\details{
This dataset was constructed by combining raw species presence data from the Gentry forest plot dataset
with phylogenetic information and cleaned taxonomic names. The species names were standardized by removing
uncertain annotations such as "cf." and "aff." and mapped to a phylogenetic tree using S.PhyloMaker
with scenario S1. Species not matched to the phylogeny (less than 2\% of counts) were removed from the
final dataset.

The data can be used for analyses involving phylogenetic diversity,
community ecology, and evolutionary relationships among species in
forest plots.
}
\examples{
library(phyloseq)
data(gentry)

# view the OTU table (species counts)
otu_table(gentry)[1:5, 1:20]

# the phylogenetic tree
phy_tree(gentry)

# view the sample data
sample_data(gentry) |> head()

# extract taxonomic classifications
tax_table(gentry) |> head()


}
\keyword{community}
\keyword{datasets}
\keyword{ecology}
\keyword{forest}
\keyword{phyloseq}
