% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpqDist.R
\name{get_avg_distances_to_set}
\alias{get_avg_distances_to_set}
\title{Average distance between each individual point and a set of base
samples}
\usage{
get_avg_distances_to_set(d_obj, base_samples, sample_data = NULL)
}
\arguments{
\item{d_obj}{A distance object describing the distances between all
the samples}

\item{base_samples}{The samples against which all the others are
measured.}

\item{sample_data}{(optional) Extra data that can be added to the
data frame containing the average distances. Useful when you
want to plot average distances against other covariates.}
}
\value{
A data frame with one column containing the average
distances between each sample and the base
samples. Specifically, for sample i, the average distance is
sum_{bs in base_samples} d(i, bs) / |base_samples|. Remaining
columns in the data frame are sample names and sample_data (if
it was provided.)
}
\description{
Average distance between each individual point and a set of base
samples
}
