% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpqDist.R
\name{get_dr}
\alias{get_dr}
\title{Compute normalized eigenvalues for the MPQr distance}
\usage{
get_dr(eig, r)
}
\arguments{
\item{eig}{An 'eigen' object. Should be obtained by calling 'eigen' on the covariance matrix obtained from a phylogenetic tree T.}

\item{r}{Numeric scalar in \eqn{[0,1]} specifying the tuning parameter of the \eqn{\MPQr} distance.}
}
\value{
A numeric vector of length \code{length(eig$values)}.
}
\description{
Given an object containing the eigendecomposition of the matrix \eqn{\T} used for the PQ distance, , this function computes the vector of normalized weights corresponding to the parameter \eqn{r \in [0,1]}.
}
\examples{
eig <- eigen(matrix(c(2,1,1,2), 2, 2))
# when r = 0, the eigenvalues are equal
mpqDist:::get_dr(eig, r = 0)
mpqDist:::get_dr(eig, r = 0.5)
# when r = 1, the eigenvalues are proportional to the original eigenvalues
mpqDist:::get_dr(eig, r = 1)
}
\keyword{internal}
