% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpqDist.R
\name{get_null_mean_and_variance}
\alias{get_null_mean_and_variance}
\title{Compute null distribution statistics for the MPQr distance}
\usage{
get_null_mean_and_variance(physeq, rvec = r_transform(0:100/100))
}
\arguments{
\item{physeq}{A \code{\link[phyloseq]{phyloseq}} object containing:
\itemize{
\item a phylogenetic tree (accessible via \code{\link[phyloseq]{phy_tree}})
\item an OTU/ASV abundance table (accessible via \code{\link[phyloseq]{otu_table}})
The abundance table is used to compute per-taxon standard deviations.
}}

\item{rvec}{A numeric vector of r values in \link{0,1} for which to
compute null distribution summaries. Defaults to \code{r_transform(0:100/100)}.}
}
\value{
A named list of numeric vectors (each of length \code{length(rvec)}). Elements of the list give some values related to the distribution of the MPQr distance between two samples under a null distribution.
Under the null distribution, each sample has independent, normally-distributed elements with standard deviation equal to the empirical standard deviation of the observed variables.
The statistics are as follows:
\itemize{
\item \code{means}: Approximate means of the MPQr distance between two samples in the independent-normal null model.
\item \code{sds}: Approximate standard deviations of the MPQr distance between two samples in the independent normal null model.
\item \code{medians}: Median of the null distribution.
\item \code{lower}: Lower 95\% confidence bound (2.5\% quantile of the null distribution)
\item \code{upper}: Upper 95\% confidence bound (97.5\% quantile of the null distribution)
}
}
\description{
For a given dataset, encoded as a phyloseq object, and a set of r
values, this function computes approximate means, standard
deviations, medians, and selected quantiles of the MPQr distance
under a null model in which sample values are independent Gaussian
random variables with the same variance as in the observed data.
}
\examples{
data(gentry)
get_null_mean_and_variance(gentry, rvec = c(0, 0.5, 1))

}
