% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpqDist.R
\name{mpq_and_sample_distances}
\alias{mpq_and_sample_distances}
\title{Creates a data frame containing MPQr and sample distances}
\usage{
mpq_and_sample_distances(d_obj, sample_distances)
}
\arguments{
\item{d_obj}{A distance object containing the MPQr distances
between the samples.}

\item{sample_distances}{A distance object containing the distances
between samples.}
}
\value{
A data frame
}
\description{
Creates a data frame containing MPQr and sample distances
}
\keyword{internal}
