% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpqDist.R
\name{plot_all_pairs}
\alias{plot_all_pairs}
\title{Creates an animated plot of MPQ distances against some other
measure of distance between sites}
\usage{
plot_all_pairs(physeq, sample_distances, return_ggplot = FALSE, ...)
}
\arguments{
\item{physeq}{A phyloseq object.}

\item{return_ggplot}{If TRUE, the function will return a ggplot
object which can be made into an animated plot by calling
ggplotly on the ggplot object. If FALSE, the function will
create the animated plot. The option of returning the ggplot
object is primarily to allow for more customization by the
user.}

\item{distance_variables}{Names of columns in sample_data(physeq)
to use to compute distances}

\item{distance_fn}{A function that takes a matrix and computes
distances between all pairs of rows. It should return an object
of class 'dist'.}
}
\value{
Either an animated plot created by \code{ggplotly} (if
\code{return_ggplot = FALSE}) or a ggplot object (if \code{return_ggplot = TRUE})
}
\description{
Creates an animated plot of MPQ distances against some other
measure of distance between sites
}
