% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpqDist.R
\name{plot_dist_to_reference_samples}
\alias{plot_dist_to_reference_samples}
\title{Makes an animated plot showing the the distances between each
sample and a set of reference samples}
\usage{
plot_dist_to_reference_samples(
  physeq,
  reference_samples,
  x_variable,
  return_ggplot = FALSE,
  ...
)
}
\arguments{
\item{physeq}{A phyloseq object containing the abundances, a
phylogenetic tree, and sample data}

\item{reference_samples}{Either the names or the indices of the set
of reference samples.}

\item{x_variable}{A variable for plotting the distances against on
the x axis. Should be a string giving the name of one of the
columns of \code{sample_data(phy_seq)}.}

\item{return_ggplot}{If TRUE, the function will return a ggplot
object which can be made into an animated plot by calling
ggplotly on the ggplot object. If FALSE, the function will
create the animated plot. The option of returning the ggplot
object is primarily to allow for more customization by the
user.}

\item{...}{Extra objects to be passed to \code{aes_string} in ggplot.}
}
\value{
Either an animated plot created by \code{ggplotly} (if
\code{return_ggplot = FALSE}) or a ggplot object (if \code{return_ggplot = TRUE})
}
\description{
Makes an animated plot showing the the distances between each
sample and a set of reference samples
}
