% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpqDist.R
\name{r_transform}
\alias{r_transform}
\title{A transformation that tends to work better than a linear grid between 0 and 1}
\usage{
r_transform(r)
}
\arguments{
\item{r}{A value between 0 and 1}
}
\value{
Another value between 0 and 1.
}
\description{
A transformation that tends to work better than a linear grid between 0 and 1
}
