% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small_otutab}
\alias{small_otutab}
\title{Subsampled OTU Table from the Gentry Dataset}
\format{
A numeric matrix with 50 columns (taxa) and rows corresponding to samples/sites.
}
\usage{
small_otutab
}
\description{
A small subset of the OTU (species abundance) table from the \code{gentry} phyloseq object,
containing counts for 50 randomly selected taxa (out of the original 585). This object
is primarily intended for use in examples and demonstrations, and is not recommended
for substantive ecological or phylogenetic analyses.
}
\details{
The subset was created by randomly sampling 50 taxa indices and subsetting
the OTU table from the \code{gentry} dataset using \code{subset_taxa()}.
The resulting OTU table was coerced to a matrix for convenience.
}
\examples{
dim(small_otutab)
## note that because we subsampled taxa, many of the sites are exactly the same and have no taxa
head(small_otutab)

}
\seealso{
\code{\link{gentry}}
}
\keyword{data}
\keyword{datasets}
\keyword{example}
