% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small_tree}
\alias{small_tree}
\title{Subsampled Phylogenetic Tree from the Gentry Dataset}
\format{
A \code{\link[ape]{phylo}} object representing the phylogenetic relationships of the
50 selected taxa.
}
\usage{
small_tree
}
\description{
A phylogenetic tree corresponding to the 50 randomly selected taxa used in \code{small_otutab},
extracted from the \code{gentry} phyloseq object. This tree is intended solely for
illustrative or testing purposes and should not be used for formal phylogenetic analysis.
}
\details{
The tree was extracted by subsetting the original phylogenetic tree in \code{gentry}
to include only the taxa corresponding to the sampled subset used in \code{small_otutab}.
}
\examples{
\dontrun{
plot(small_tree)
}

}
\seealso{
\code{\link{gentry}}, \code{\link{small_otutab}}
}
\keyword{data}
\keyword{datasets}
\keyword{example}
\keyword{phylogenetics}
