% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpqDist.R
\name{make_animation_df}
\alias{make_animation_df}
\title{Sets up a data frame for plotly to make an animation}
\usage{
make_animation_df(mpq_distances, fn, means_and_vars = NULL, ...)
}
\arguments{
\item{fn}{A function to apply to each of the distance objects that
creates a data frame with the variables in the animation. Its
first argument should be a distance object. For example,
\code{get_avg_distances_to_set}.}

\item{...}{Extra arguments that are passed to fn.}

\item{distance_list}{A list of distance objects, for example, the
\code{distance} element from the output of
\code{get_mpq_distances}.}
}
\value{
A data frame that has one column called \code{frame} (which
can be used by ggplot2 and plotly to make an animation).
}
\description{
Sets up a data frame for plotly to make an animation
}
