% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpqDist.R
\name{plot_all_pairs}
\alias{plot_all_pairs}
\title{Creates an animated plot of MPQ distances against some other
measure of distance between sites}
\usage{
plot_all_pairs(physeq, sample_distances, return_ggplot = FALSE)
}
\arguments{
\item{physeq}{A phyloseq object containing species abundances, a
phylogenetic tree, and sample data.}

\item{sample_distances}{Either a distance object or a (n_samples x
n_samples) matrix containing distances between the samples.}

\item{return_ggplot}{If TRUE, the function will return a ggplot
object which can be made into an animated plot by calling
ggplotly on it. If FALSE, the function will create the animated
plot. The option of returning the ggplot object is primarily to
allow for more customization by the user.}
}
\value{
Either an animated plot created by \code{ggplotly} (if
\code{return_ggplot = FALSE}) or a ggplot object (if \code{return_ggplot = TRUE})
}
\description{
Creates an animated plot of MPQ distances against some other
measure of distance between sites
}
\examples{
data(gentry)
gentry_small = subset_samples(gentry, Precip > 4000)
longitude_distance = dist(sample_data(gentry_small)$Long)
plot_all_pairs(gentry_small, longitude_distance)
}
